clarify tier boundary and split qlinear pooling
